Changed Charge get flow into the new requset flow
